+++ contrib/nx-oxcaml/lib/op_scatter.ml
@@ -57,15 +57,14 @@ let scatter_float64 mode (src : float# array) (dst : float# array) ishape dshape
       let idx =
         Int32.to_int (Int32_u.to_int32 (Array.unsafe_get idx_arr !idx_lin))
       in
-      if idx < 0 || idx >= Array.unsafe_get dshape axis then
-        invalid_arg "scatter: index out of bounds";
-      let dst_lin = !dst_base + (idx * axis_stride) in
-      let v = Array.unsafe_get src !src_lin in
-      (match mode with
-      | `Set -> Array.unsafe_set dst dst_lin v
-      | `Add ->
-          Array.unsafe_set dst dst_lin
-            (Float_u.add (Array.unsafe_get dst dst_lin) v))
+      if idx >= 0 && idx < Array.unsafe_get dshape axis then (
+        let dst_lin = !dst_base + (idx * axis_stride) in
+        let v = Array.unsafe_get src !src_lin in
+        match mode with
+        | `Set -> Array.unsafe_set dst dst_lin v
+        | `Add ->
+            Array.unsafe_set dst dst_lin
+              (Float_u.add (Array.unsafe_get dst dst_lin) v))
     in
     let i = ref start_idx in
     let n4 = end_idx - 3 in
@@ -107,15 +106,14 @@ let scatter_float32 mode (src : float32# array) (dst : float32# array) ishape
       let idx =
         Int32.to_int (Int32_u.to_int32 (Array.unsafe_get idx_arr !idx_lin))
       in
-      if idx < 0 || idx >= Array.unsafe_get dshape axis then
-        invalid_arg "scatter: index out of bounds";
-      let dst_lin = !dst_base + (idx * axis_stride) in
-      let v = Array.unsafe_get src !src_lin in
-      (match mode with
-      | `Set -> Array.unsafe_set dst dst_lin v
-      | `Add ->
-          Array.unsafe_set dst dst_lin
-            (Float32_u.add (Array.unsafe_get dst dst_lin) v))
+      if idx >= 0 && idx < Array.unsafe_get dshape axis then (
+        let dst_lin = !dst_base + (idx * axis_stride) in
+        let v = Array.unsafe_get src !src_lin in
+        match mode with
+        | `Set -> Array.unsafe_set dst dst_lin v
+        | `Add ->
+            Array.unsafe_set dst dst_lin
+              (Float32_u.add (Array.unsafe_get dst dst_lin) v))
     in
     let i = ref start_idx in
     let n4 = end_idx - 3 in
@@ -157,15 +155,14 @@ let scatter_int8 mode (src : int8# array) (dst : int8# array) ishape dshape axis
       let idx =
         Int32.to_int (Int32_u.to_int32 (Array.unsafe_get idx_arr !idx_lin))
       in
-      if idx < 0 || idx >= Array.unsafe_get dshape axis then
-        invalid_arg "scatter: index out of bounds";
-      let dst_lin = !dst_base + (idx * axis_stride) in
-      let v = Array.unsafe_get src !src_lin in
-      (match mode with
-      | `Set -> Array.unsafe_set dst dst_lin v
-      | `Add ->
-          Array.unsafe_set dst dst_lin
-            (Int8_u.add (Array.unsafe_get dst dst_lin) v))
+      if idx >= 0 && idx < Array.unsafe_get dshape axis then (
+        let dst_lin = !dst_base + (idx * axis_stride) in
+        let v = Array.unsafe_get src !src_lin in
+        match mode with
+        | `Set -> Array.unsafe_set dst dst_lin v
+        | `Add ->
+            Array.unsafe_set dst dst_lin
+              (Int8_u.add (Array.unsafe_get dst dst_lin) v))
     in
     let i = ref start_idx in
     let n4 = end_idx - 3 in
@@ -207,15 +204,14 @@ let scatter_int16 mode (src : int16# array) (dst : int16# array) ishape dshape
       let idx =
         Int32.to_int (Int32_u.to_int32 (Array.unsafe_get idx_arr !idx_lin))
       in
-      if idx < 0 || idx >= Array.unsafe_get dshape axis then
-        invalid_arg "scatter: index out of bounds";
-      let dst_lin = !dst_base + (idx * axis_stride) in
-      let v = Array.unsafe_get src !src_lin in
-      (match mode with
-      | `Set -> Array.unsafe_set dst dst_lin v
-      | `Add ->
-          Array.unsafe_set dst dst_lin
-            (Int16_u.add (Array.unsafe_get dst dst_lin) v))
+      if idx >= 0 && idx < Array.unsafe_get dshape axis then (
+        let dst_lin = !dst_base + (idx * axis_stride) in
+        let v = Array.unsafe_get src !src_lin in
+        match mode with
+        | `Set -> Array.unsafe_set dst dst_lin v
+        | `Add ->
+            Array.unsafe_set dst dst_lin
+              (Int16_u.add (Array.unsafe_get dst dst_lin) v))
     in
     let i = ref start_idx in
     let n4 = end_idx - 3 in
@@ -257,15 +253,14 @@ let scatter_int32 mode (src : int32# array) (dst : int32# array) ishape dshape
       let idx =
         Int32.to_int (Int32_u.to_int32 (Array.unsafe_get idx_arr !idx_lin))
       in
-      if idx < 0 || idx >= Array.unsafe_get dshape axis then
-        invalid_arg "scatter: index out of bounds";
-      let dst_lin = !dst_base + (idx * axis_stride) in
-      let v = Array.unsafe_get src !src_lin in
-      (match mode with
-      | `Set -> Array.unsafe_set dst dst_lin v
-      | `Add ->
-          Array.unsafe_set dst dst_lin
-            (Int32_u.add (Array.unsafe_get dst dst_lin) v))
+      if idx >= 0 && idx < Array.unsafe_get dshape axis then (
+        let dst_lin = !dst_base + (idx * axis_stride) in
+        let v = Array.unsafe_get src !src_lin in
+        match mode with
+        | `Set -> Array.unsafe_set dst dst_lin v
+        | `Add ->
+            Array.unsafe_set dst dst_lin
+              (Int32_u.add (Array.unsafe_get dst dst_lin) v))
     in
     let i = ref start_idx in
     let n4 = end_idx - 3 in
@@ -307,15 +302,14 @@ let scatter_int64 mode (src : int64# array) (dst : int64# array) ishape dshape
       let idx =
         Int32.to_int (Int32_u.to_int32 (Array.unsafe_get idx_arr !idx_lin))
       in
-      if idx < 0 || idx >= Array.unsafe_get dshape axis then
-        invalid_arg "scatter: index out of bounds";
-      let dst_lin = !dst_base + (idx * axis_stride) in
-      let v = Array.unsafe_get src !src_lin in
-      (match mode with
-      | `Set -> Array.unsafe_set dst dst_lin v
-      | `Add ->
-          Array.unsafe_set dst dst_lin
-            (Int64_u.add (Array.unsafe_get dst dst_lin) v))
+      if idx >= 0 && idx < Array.unsafe_get dshape axis then (
+        let dst_lin = !dst_base + (idx * axis_stride) in
+        let v = Array.unsafe_get src !src_lin in
+        match mode with
+        | `Set -> Array.unsafe_set dst dst_lin v
+        | `Add ->
+            Array.unsafe_set dst dst_lin
+              (Int64_u.add (Array.unsafe_get dst dst_lin) v))
     in
     let i = ref start_idx in
     let n4 = end_idx - 3 in
@@ -357,13 +351,12 @@ let scatter_bool mode (src : bool array) (dst : bool array) ishape dshape axis
       let idx =
         Int32.to_int (Int32_u.to_int32 (Array.unsafe_get idx_arr !idx_lin))
       in
-      if idx < 0 || idx >= Array.unsafe_get dshape axis then
-        invalid_arg "scatter: index out of bounds";
-      let dst_lin = !dst_base + (idx * axis_stride) in
-      let v = Array.unsafe_get src !src_lin in
-      (match mode with
-      | `Set -> Array.unsafe_set dst dst_lin v
-      | `Add -> Array.unsafe_set dst dst_lin (Array.unsafe_get dst dst_lin || v))
+      if idx >= 0 && idx < Array.unsafe_get dshape axis then (
+        let dst_lin = !dst_base + (idx * axis_stride) in
+        let v = Array.unsafe_get src !src_lin in
+        match mode with
+        | `Set -> Array.unsafe_set dst dst_lin v
+        | `Add -> Array.unsafe_set dst dst_lin (Array.unsafe_get dst dst_lin || v))
     in
     let i = ref start_idx in
     let n4 = end_idx - 3 in